WIP display in proper cell - adapt unit tests
